Modified the build machinery so that we can build and link to Trilinos 10 as well as Trilinos 9. This is still not as nice as I would like because I haven't invested enough time into the autotools part. The script in external_distributions/trilinos will detect any file of the form trilinos*.tar.gz and extract the major version number to do the right thing (autotools vs cmake). I could not find an easy way of making autoconf use wildcards in the same way, so have hard-coded the .tar.gz files to be trilinos-9.0.2.tar.gz (as before) and trilinos-10.8.5-Source.tar.gz (the most current release as of 7th Feb 2012).
Needed one change in trilinos_solver.h. The latest version of Trilinos's
ML preconditioner uses the communicator of the matrix in its destructor, so
the preconditioner must be deleted before the matrix.
All tests pass on my machine with Trilinos 9 or Trilinos 10.

Also added LAPACK files clarnd.f and zlarnd.f that were 
required by Trilinos 10

At the same time, I've modified the double_vector_with_halo.h/.cc so that
the Halo/Haloed_equations are unsigned and passed between processors with
an MPI_UNSIGNED variable. Hopefully this is now robust.

I note that I haven't checked all possible configurations of 
Trilinos/BLAS/LAPACK, etc, so there may be an error, but I have tried to
mirror what was done with Trilinos 9, so I *hope* it'll all be OK. If not
feel free to shout!



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@264 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/generic/double_vector_with_halo.cc
@@ -125,7 +125,7 @@ namespace oomph
    //Prepare the data to be sent
    //Always resize to at least one
    if(send_data_count==0) {++send_data_count;}
-   Vector<int> send_data(send_data_count);
+   Vector<unsigned> send_data(send_data_count);
    //Iterate over the entries in the map
    unsigned count=0;
    for(std::map<unsigned,Vector<unsigned> >::iterator it
@@ -135,7 +135,7 @@ namespace oomph
     for(Vector<unsigned>::iterator it2 = it->second.begin();
         it2 != it->second.end();++it2)
      {
-      send_data[count] = static_cast<int>(*it2);
+      send_data[count] = (*it2);
       ++count;
      }
    }
@@ -157,10 +157,10 @@ namespace oomph
    Haloed_eqns.resize(receive_data_count);
    //Send the data between all the processors
    MPI_Alltoallv(&send_data[0],&send_n[0],&send_displacement[0],
-                 MPI_INT,
+                 MPI_UNSIGNED,
                  &Haloed_eqns[0],&Haloed_n[0],
                  &Haloed_displacement[0],
-                 MPI_INT,
+                 MPI_UNSIGNED,
                  dist_pt->communicator_pt()->mpi_comm());
    
    //Finally, we translate the map of halo entries into the permanent
